bicomponent.dist {sna}	R Documentation

Calculate the Bicomponents of a Graph

Description

bicomponent.dist returns the bicomponents of an input graph, along with size distribution and membership information.

Usage

bicomponent.dist(dat, symmetrize = c("strong", "weak"))

Arguments

dat
a graph or graph stack.

symmetrize
symmetrization rule to apply when pre-processing the input (see symmetrize).

Details

The bicomponents of undirected graph G are its maximal 2-connected vertex sets. bicomponent.dist calculates the bicomponents of G, after first coercing to undirected form using the symmetrization rule in symmetrize. In addition to bicomponent memberships, various summary statistics regarding the bicomponent distribution are returned; see below.

Value

A list containing

members
A list, with one entry per bicomponent, containing component members.

memberships
A vector of component memberships, by vertex. (Note: memberships may not be unique.) Vertices not belonging to any bicomponent have membership values of NA.

csize
A vector of component sizes, by bicomponent.

cdist
A vector of length |V(G)| with the (unnormalized) empirical distribution function of bicomponent sizes.

Note

Remember that bicomponents can intersect; when this occurs, the relevant vertices' entries in the membership vector are assigned to one of the overlapping bicomponents on an arbitrary basis. The members element of the return list is the safe way to recover membership information.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Brandes, U. and Erlebach, T. (2005). Network Analysis: Methodological Foundations. Berlin: Springer.

See Also

component.dist, cutpoints, symmetrize

Examples

#Draw a moderately sparse graph
g<-rgraph(25,tp=2/24,mode="graph")

#Compute the bicomponents
bicomponent.dist(g)

[Package sna version 2.8 Index]