brokerage {sna}	R Documentation

Perform a Gould-Fernandez Brokerage Analysis

Description

Performs the brokerage analysis of Gould and Fernandez on one or more input graphs, given a class membership vector.

Usage

brokerage(g, cl)

Arguments

g
one or more input graphs.

cl
a vector of class memberships.

Details

Gould and Fernandez (following Marsden and others) describe brokerage as the role played by a social actor who mediates contact between two alters. More formally, vertex v is a broker for distinct vertices a and b iff a -> v -> b and a -!> b. Where actors belong to a priori distinct groups, group membership may be used to segment brokerage roles into particular types. Let A -> B -> C denote the two-path associated with a brokerage structure, such that some vertex from group B brokers the connection from some vertex from group A to a vertex in group C. The types of brokerage roles defined by Gould and Fernandez (and their accompanying two-path structures) are then defined in terms of group membership as follows:

w_I: Coordinator role; the broker mediates contact between two individuals from his or her own group. Two-path structure: A -> A -> A

w_O: Itinerant broker role; the broker mediates contact between two individuals from a single group to which he or she does not belong. Two-path structure: A -> B -> A

b_{OI}: Gatekeeper role; the broker mediates an incoming contact from an out-group member to an in-group member. Two-path structure: A -> B -> B

b_{IO}: Representative role; the broker mediates an outgoing contact from an in-group member to an out-group member. Two-path structure: A -> A -> B

b_O: Liaison role; the broker mediates contact between two individuals from different groups, neither of which is the group to which he or she belongs. Two-path structure: A -> B -> C

t: Total (cumulative) brokerage role occupancy. (Any of the above two-paths.)

The brokerage score for a given vertex with respect to a given role is the number of ordered pairs having the appropriate group membership(s) brokered by said vertex. brokerage computes the brokerage scores for each vertex, given an input graph and vector of class memberships. Aggregate scores are also computed at the graph level, which correspond to the total frequency of each role type within the network structure. Expectations and variances of the brokerage scores conditional on size and density are computed, along with approximate z-tests for incidence of brokerage. (Note that the accuracy of the normality assumption is not known in the general case; see Gould and Fernandez (1989) for details. Simulation-based tests may be desirable as an alternative.)

Value

An object of class brokerage, containing the following elements:

raw.nli
The matrix of observed brokerage scores, by vertex

exp.nli
The matrix of expected brokerage scores, by vertex

sd.nli
The matrix of predicted brokerage score standard deviations, by vertex

z.nli
The matrix of standardized brokerage scores, by vertex

raw.gli
The vector of observed aggregate brokerage scores

exp.gli
The vector of expected aggregate brokerage scores

sd.gli
The vector of predicted aggregate brokerage score standard deviations

z.gli
The vector of standardized aggregate brokerage scores

exp.grp
The matrix of expected brokerage scores, by group

sd.grp
The matrix of predicted brokerage score standard deviations, by group

cl
The vector of class memberships

clid
The original class names

n
The input class sizes

N
The order of the input network

Author(s)

Carter T. Butts buttsc@uci.edu

References

Gould, R.V. and Fernandez, R.M. 1989. “Structures of Mediation: A Formal Approach to Brokerage in Transaction Networks.” Sociological Methodology, 19: 89-126.

See Also

triad.census, gtrans

Examples

#Draw a random network with 3 groups
g<-rgraph(15)
cl<-rep(1:3,5)

#Compute a brokerage object
b<-brokerage(g,cl)
summary(b)

[Package sna version 2.8 Index]