centralization {sna}	R Documentation

Find the Centralization of a Given Network, for Some Measure of Centrality

Description

Centralization returns the centralization GLI (graph-level index) for a given graph in dat, given a (node) centrality measure FUN. Centralization follows Freeman's (1979) generalized definition of network centralization, and can be used with any properly defined centrality measure. This measure must be implemented separately; see the references below for examples.

Usage

centralization(dat, FUN, g=NULL, mode="digraph", diag=FALSE, 
    normalize=TRUE, ...)

Arguments

dat
one or more input graphs.

FUN
Function to return nodal centrality scores.

g
Integer indicating the index of the graph for which centralization should be computed. By default, all graphs are employed.

mode
String indicating the type of graph being evaluated. "digraph" indicates that edges should be interpreted as directed; "graph" indicates that edges are undirected. mode is set to "digraph" by default.

diag
Boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

normalize
Boolean indicating whether or not the centralization score should be normalized to the theoretical maximum. (Note that this function relies on FUN to return this value when called with tmaxdev==TRUE.) By default, tmaxdev==TRUE.

...
Additional arguments to FUN.

Details

The centralization of a graph G for centrality measure C(v) is defined (as per Freeman (1979)) to be:

C^*(G) = sum( |max(C(v))-C(i)|, i in V(G) )

Or, equivalently, the absolute deviation from the maximum of C on G. Generally, this value is normalized by the theoretical maximum centralization score, conditional on |V(G)|. (Here, this functionality is activated by normalize.) Centralization depends on the function specified by FUN to return the vector of nodal centralities when called with dat and g, and to return the theoretical maximum value when called with the above and tmaxdev==TRUE. For an example of such a centrality routine, see degree.

Value

The centralization of the specified graph.

Note

See cugtest for null hypothesis tests involving centralization scores.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Freeman, L.C. (1979). “Centrality in Social Networks I: Conceptual Clarification.” Social Networks, 1, 215-239.

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

cugtest

Examples

#Generate some random graphs
dat<-rgraph(5,10)
#How centralized is the third one on indegree?
centralization(dat,g=3,degree,cmode="indegree")
#How about on total (Freeman) degree?
centralization(dat,g=3,degree)

[Package sna version 2.8 Index]