coleman {sna}	R Documentation

Coleman's High School Friendship Data

Description

James Coleman (1964) reports research on self-reported friendship ties among 73 boys in a small high school in Illinois over the 1957-1958 academic year. Networks of reported ties for all 73 informants are provided for two time points (fall and spring).

Usage

data(coleman)

Format

An adjacency array containing two directed, unvalued 73-node networks:

[1,,] 	 Fall 	 binary matrix 	 Friendship for Fall, 1957

[2,,] 	 Spring 	 binary matrix 	 Friendship for Spring, 1958

Details

Both networks reflect answers to the question, “What fellows here in school do you go around with most often?” with the presence of an (i,j,k) edge indicating that j nominated k in time period i. The data are unvalued and directed; although the self-reported ties are highly reciprocal, unreciprocated nominations are possible.

It should be noted that, although this data is usually described as “friendship,” the sociometric item employed might be more accurately characterized as eliciting “frequent elective interaction.” This should be borne in mind when interpreting this data.

References

Coleman, J. S. (1964). Introduction to Mathermatical Sociology. New York: Free Press.

Examples

data(coleman)

#Plot showing edges by time point
gplot(coleman[1,,]|coleman[2,,],edge.col=2*coleman[1,,]+3*coleman[2,,])

[Package sna version 2.8 Index]