components {sna}	R Documentation

Find the Number of (Maximal) Components Within a Given Graph

Description

Returns the number of components within dat, using the connectedness rule given in connected.

Usage

components(dat, connected="strong", comp.dist.precomp=NULL)

Arguments

dat
one or more input graphs.

connected
the the component definition to be used by component.dist during component extraction.

comp.dist.precomp
a component size distribution object from component.dist (optional).

Details

The connected parameter corresponds to the rule parameter of component.dist. By default, components returns the number of strong components, but other component types can be returned if so desired. (See component.dist for details.) For symmetric matrices, this is obviously a moot point.

Value

A vector containing the number of components for each graph in dat

Author(s)

Carter T. Butts buttsc@uci.edu

References

West, D.B. (1996). Introduction to Graph Theory. Upper Saddle River, NJ: Prentice Hall.

See Also

component.dist, symmetrize

Examples

g<-rgraph(20,tprob=0.05)   #Generate a sparse random graph

#Find weak components
components(g,connected="weak")

#Find strong components
components(g,connected="strong")

[Package sna version 2.8 Index]