connectedness {sna}	R Documentation

Compute Graph Connectedness Scores

Description

connectedness takes one or more graphs (dat) and returns the Krackhardt connectedness scores for the graphs selected by g.

Usage

connectedness(dat, g=NULL)

Arguments

dat
one or more graphs.

g
index values for the graphs to be utilized; by default, all graphs are selected.

Details

Krackhardt's connectedness for a digraph G is equal to the fraction of all dyads, \{i,j\}, such that there exists an undirected path from i to j in G. (This, in turn, is just the density of the weak reachability graph of G.) Obviously, the connectedness score ranges from 0 (for the null graph) to 1 (for weakly connected graphs).

Connectedness is one of four measures (connectedness, efficiency, hierarchy, and lubness) suggested by Krackhardt for summarizing hierarchical structures. Each corresponds to one of four axioms which are necessary and sufficient for the structure in question to be an outtree; thus, the measures will be equal to 1 for a given graph iff that graph is an outtree. Deviations from unity can be interpreted in terms of failure to satisfy one or more of the outtree conditions, information which may be useful in classifying its structural properties.

Value

A vector containing the connectedness scores

Note

The four Krackhardt indices are, in general, nondegenerate for a relatively narrow band of size/density combinations (efficiency being the sole exception). This is primarily due to their dependence on the reachability graph, which tends to become complete rapidly as size/density increase. See Krackhardt (1994) for a useful simulation study.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Krackhardt, David. (1994). “Graph Theoretical Dimensions of Informal Organizations.” In K. M. Carley and M. J. Prietula (Eds.), Computational Organization Theory, 89-111. Hillsdale, NJ: Lawrence Erlbaum and Associates.

See Also

connectedness, efficiency, hierarchy, lubness, reachability

Examples

#Get connectedness scores for graphs of varying densities
connectedness(rgraph(10,5,tprob=c(0.1,0.25,0.5,0.75,0.9)))

[Package sna version 2.8 Index]