dyad.census {sna}	R Documentation

Compute a Holland and Leinhardt MAN Dyad Census

Description

dyad.census computes a Holland and Leinhardt dyad census on the graphs of dat selected by g.

Usage

dyad.census(dat, g=NULL)

Arguments

dat
one or more graphs.

g
the elements of dat to be included; by default, all graphs are processed.

Details

Each dyad in a directed graph may be in one of four states: the null state (a <-!-> b), the complete or mutual state (a <--> b), and either of two asymmetric states (a --> b or a <-- b). Holland and Leinhardt's dyad census classifies each dyad into the mutual, asymmetric, or null categories, counting the number of each within the digraph. These counts can be used as the basis for null hypothesis tests (since their distributions are known under assumptions such as constant edge probability), or for the generation of random graphs (e.g., via the U|MAN distribution, which conditions on the numbers of mutual, asymmetric, and null dyads in each graph).

Value

A matrix whose three columns contain the counts of mutual, asymmetric, and null dyads (respectively) for each graph

Author(s)

Carter T. Butts buttsc@uci.edu

References

Holland, P.W. and Leinhardt, S. (1970). “A Method for Detecting Structure in Sociometric Data.” American Journal of Sociology, 76, 492-513.

Wasserman, S., and Faust, K. (1994). “Social Network Analysis: Methods and Applications.” Cambridge: Cambridge University Press.

See Also

mutuality, grecip, rguman triad.census, kcycle.census, kpath.census

Examples

#Generate a dyad census of random data with varying densities
dyad.census(rgraph(15,5,tprob=c(0.1,0.25,0.5,0.75,0.9)))

[Package sna version 2.8 Index]