event2dichot {sna}	R Documentation

Convert an Observed Event Matrix to a Dichotomous matrix

Description

Given one or more valued adjacency matrices (possibly derived from observed interaction “events”), event2dichot returns dichotomized equivalents.

Usage

event2dichot(m, method="quantile", thresh=0.5, leq=FALSE)

Arguments

m
one or more (valued) input graphs.

method
one of “quantile,” “rquantile,” “cquantile,” “mean,” “rmean,” “cmean,” “absolute,” “rank,” “rrank,” or “crank”.

thresh
dichotomization thresholds for ranks or quantiles.

leq
boolean indicating whether values less than or equal to the threshold should be taken as existing edges; the alternative is to use values strictly greater than the threshold.

Details

The methods used for choosing dichotomization thresholds are as follows:

quantile: specified quantile over the distribution of all edge values

rquantile: specified quantile by row

cquantile: specified quantile by column

mean: grand mean

rmean: row mean

cmean: column mean

absolute: the value of thresh itself

rank: specified rank over the distribution of all edge values

rrank: specified rank by row

crank: specified rank by column

Note that when a quantile, rank, or value is said to be “specified,” this refers to the value of thresh.

Value

The dichotomized data matrix (or matrices)

Author(s)

Carter T. Butts buttsc@uci.edu

References

Wasserman, S. and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

Examples

#Draw a matrix of normal values
n<-matrix(rnorm(25),nrow=5,ncol=5)

#Dichotomize by the mean value
event2dichot(n,"mean")

#Dichotomize by the 0.95 quantile
event2dichot(n,"quantile",0.95)

[Package sna version 2.8 Index]