gclust.centralgraph {sna}	R Documentation

Get Central Graphs Associated with Graph Clusters

Description

Calculates central graphs associated with particular graph clusters (as indicated by the k partition of h).

Usage

gclust.centralgraph(h, k, dat, ...)

Arguments

h
an hclust object, based on a graph stack in dat.

k
the number of groups to evaluate.

dat
one or more graphs (on which the clustering was performed).

...
additional arguments to centralgraph.

Details

gclust.centralgraph uses cutree to cut the hierarchical clustering in h into k groups. centralgraph is then called on each cluster, and the results are returned as a graph stack. This is a useful tool for interpreting clusters of (labeled) graphs, with the resulting central graphs being subsequently analyzed using standard SNA methods.

Value

An array containing the stack of central graph adjacency matrices

Author(s)

Carter T. Butts buttsc@uci.edu

References

Butts, C.T., and Carley, K.M. (2001). “Multivariate Methods for Interstructural Analysis.” CASOS working paper, Carnegie Mellon University.

See Also

hclust, centralgraph, gclust.boxstats, gdist.plotdiff, gdist.plotstats

Examples

#Create some random graphs
g<-rgraph(10,20,tprob=c(rbeta(10,15,2),rbeta(10,2,15)))

#Find the Hamming distances between them
g.h<-hdist(g)

#Cluster the graphs via their Hamming distances
g.c<-hclust(as.dist(g.h))

#Now find central graphs by cluster for a two cluster solution
g.cg<-gclust.centralgraph(g.c,2,g)

#Plot the central graphs
gplot(g.cg[1,,])
gplot(g.cg[2,,])

[Package sna version 2.8 Index]