gilschmidt {sna}	R Documentation

Compute the Gil-Schmidt Power Index

Description

gilschmidt computes the Gil-Schmidt Power Index for all nodes in dat, with or without normalization.

Usage

gilschmidt(dat, g = 1, nodes = NULL, gmode = "digraph", diag = FALSE, 
    tmaxdev = FALSE, normalize = TRUE)

Arguments

dat
one or more input graphs (for best performance, sna edgelists or network objects are suggested).

g
integer indicating the index of the graph for which centralities are to be calculated (or a vector thereof). By default, g=1.

nodes
list indicating which nodes are to be included in the calculation. By default, all nodes are included.

gmode
string indicating the type of graph being evaluated. "digraph" indicates that edges should be interpreted as directed; "graph" indicates that edges are undirected. gmode is set to "digraph" by default.

diag
boolean indicating whether or not the diagonal should be treated as valid data. (This has no effect on this index, but is included for compatibility with centralization.

tmaxdev
boolean indicating whether or not the theoretical maximum absolute deviation from the maximum nodal centrality should be returned. By default, tmaxdev==FALSE.

normalize
logical; should the index scores be normalized?

Details

For graph G=(V,E), let R(v,G) be the set of vertices reachable by v in V \ v. Then the Gil-Schmidt power index is defined as

C_GS(v) = sum( 1/d(v,i), i in R(v,G) )/|R(v,G)|,

where d(v,i) is the geodesic distance from v to i in G; the index is taken to be 0 for isolates. The measure takes a value of 1 when v is adjacent to all reachable vertices, and approaches 0 as the distance from v to each vertex approaches infinity. (For finite N=|V|, the minimum value is 0 if v is an isolate, and otherwise 1/(N-1).)

If normalize=FALSE is selected, then normalization by |R(v,G)| is not performed. This measure has been proposed as a better-behaved alternative to closeness (to which it is closely related).

The closeness function in the sna library can also be used to compute this index.

Value

A vector of centrality scores.

Author(s)

Carter T. Butts, buttsc@uci.edu

References

Gil, J. and Schmidt, S. (1996). “The Origin of the Mexican Network of Power”. Proceedings of the International Social Network Conference, Charleston, SC, 22-25.

Sinclair, P.A. (2009). “Network Centralization with the Gil Schmidt Power Centrality Index” Social Networks, 29, 81-92.

See Also

closeness, centralization

Examples

data(coleman)  #Load Coleman friendship network
gs<-gilschmidt(coleman,g=1:2)  #Compute the Gil-Schmidt index

#Plot G-S values in the fall, versus spring
plot(gs,xlab="Fall",ylab="Spring",main="G-S Index")
abline(0,1)

[Package sna version 2.8 Index]