gliop {sna}	R Documentation

Return a Binary Operation on GLI Values Computed on Two Graphs

Description

gliop is a wrapper which allows for an arbitrary binary operation on GLIs to be treated as a single call. This is particularly useful for test routines such as cugtest and qaptest.

Usage

gliop(dat, GFUN, OP="-", g1=1, g2=2, ...)

Arguments

dat
a collection of graphs.

GFUN
a function taking single graphs as input.

OP
the operator to use on the output of GFUN.

g1
the index of the first input graph.

g2
the index of the second input graph.

...
Additional arguments to GFUN

Details

gliop operates by evaluating GFUN on the graphs indexed by g1 and g2 and returning the result of OP as applied to the GFUN output.

Value

OP(GFUN(dat[g1, , ],...),GFUN(dat[g2, , ],...))

Note

If the output of GFUN is not sufficiently well-behaved, undefined behavior may occur. Common sense is advised.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Anderson, B.S.; Butts, C.T.; and Carley, K.M. (1999). “The Interaction of Size and Density with Graph-Level Indices.” Social Networks, 21(3), 239-267.

See Also

cugtest, qaptest

Examples

#Draw two random graphs
g<-rgraph(10,2,tprob=c(0.2,0.5))

#What is their difference in density?
gliop(g,gden,"-",1,2)

[Package sna version 2.8 Index]