gplot.target {sna}	R Documentation

Display a Graph in Target Diagram Form

Description

Displays an input graph (and associated vector) as a "target diagram," with vertices restricted to lie at fixed radii from the origin. Such displays are useful ways of representing vertex characteristics and/or local structural properties for graphs of small to medium size.

Usage

gplot.target(dat, x, circ.rad = (1:10)/10, circ.col = "blue", 
    circ.lwd = 1, circ.lty = 3, circ.lab = TRUE, circ.lab.cex = 0.75,
    circ.lab.theta = pi, circ.lab.col = 1, circ.lab.digits = 1,
    circ.lab.offset = 0.025, periph.outside = FALSE,
    periph.outside.offset = 1.2, ...)

Arguments

dat
an input graph.

x
a vector of vertex properties to be plotted (must match the dimensions of dat).

circ.rad
radii at which to draw reference circles.

circ.col
reference circle color.

circ.lwd
reference circle line width.

circ.lty
reference circle line type.

circ.lab
boolean; should circle labels be displayed?

circ.lab.cex
expansion factor for circle labels.

circ.lab.theta
angle at which to draw circle labels.

circ.lab.col
color for circle labels.

circ.lab.digits
digits to display for circle labels.

circ.lab.offset
offset for circle labels.

periph.outside
boolean; should "peripheral" vertices be drawn together beyond the normal vertex radius?

periph.outside.offset
radius at which "peripheral" vertices should be drawn if periph.outside==TRUE.

...
additional arguments to gplot.

Details

gplot.target is a front-end to gplot which implements the target diagram layout of Brandes et al. (2003). This layout seeks to optimize various aesthetic criteria, given the constraint that all vertices lie at fixed radii from the origin (set by x). One important feature of this algorithm is that vertices which belong to mutual dyads (described by Brandes et al. as “core” vertices) are treated differently from vertices which do not (“peripheral” vertices). Layout is optimized for core vertices prior to placing peripheral vertices; thus, the result may be misleading if mutuality is not a salient characteristic of the data.

The layout for gplot.target is handled by gplot.layout.target; additional parameters are specied on the associated manual page. Standard arguments may be passed to gplot, as well.

Value

A two-column matrix of vertex positions (generated by gplot.layout.target)

Author(s)

Carter T. Butts buttsc@uci.edu

References

Brandes, U.; Kenis, P.; and Wagner, D. (2003). “Communicating Centrality in Policy Network Drawings.” IEEE Transactions on Visualization and Computer Graphics, 9(2):241-253.

See Also

gplot.layout.target, gplot

Examples

#Generate a random graph
g<-rgraph(15)

#Produce a target diagram, centering by betweenness
gplot.target(g,betweenness(g))

[Package sna version 2.8 Index]