gplot3d.arrow {sna}	R Documentation

Add Arrows a Three-Dimensional Plot

Description

gplot3d.arrow draws an arrow between two pairs of points.

Usage

gplot3d.arrow(a, b, radius, color = "white", alpha = 1)

Arguments

a
a vector or three-column matrix containing origin X,Y,Z coordinates.

b
a vector or three-column matrix containing origin X,Y,Z coordinates.

radius
the arrow radius, in current plotting units. May be a vector, if multiple arrows are to be drawn.

color
the arrow color. May be a vector, if multiple arrows are being drawn.

alpha
alpha (transparency) value(s) for arrows. (May be a vector.)

Details

gplot3d.arrow draws one or more three-dimensional “arrows” from the points given in a to those given in b. Note that the “arrows” are really cones, narrowing in the direction of the destination point.

Value

None.

Author(s)

Carter T. Butts buttsc@uci.edu

See Also

gplot3d, gplot3d.loop

[Package sna version 2.8 Index]