gplot3d.loop {sna}	R Documentation

Add Loops to a Three-Dimensional Plot

Description

gplot3d.loop draws a "loop" at a specified location; this is used to designate self-ties in gplot3d.

Usage

gplot3d.loop(a, radius, color = "white", alpha = 1)

Arguments

a
a vector or three-column matrix containing origin X,Y,Z coordinates.

radius
the loop radius, in current plotting units. May be a vector, if multiple loops are to be drawn.

color
the loop color. May be a vector, if multiple loops are being drawn.

alpha
alpha (transparency) value(s) for loops. (May be a vector.)

Details

gplot3d.loop is the companion to gplot3d.arrow. The "loops" produced by this routine currently look less like loops than like "hats" – they are noticable as spike-like structures which protrude from vertices. Eventually, something more attractice will be produced by this routine.

Value

None.

Author(s)

Carter T. Butts buttsc@uci.edu

See Also

gplot3d.arrow, gplot3d, rgl-package

[Package sna version 2.8 Index]