infocent {sna}	R Documentation

Find Information Centrality Scores of Network Positions

Description

infocent takes one or more graphs (dat) and returns the information centralities of positions (selected by nodes) within the graphs indicated by g. This function is compatible with centralization, and will return the theoretical maximum absolute deviation (from maximum) conditional on size (which is used by centralization to normalize the observed centralization score).

Usage

infocent(dat, g=1, nodes=NULL, gmode="digraph", diag=FALSE, 
    cmode="weak", tmaxdev=FALSE, rescale=FALSE,tol=1e-20)

Arguments

dat
one or more input graphs.

g
integer indicating the index of the graph for which centralities are to be calculated (or a vector thereof). By default, g==1.

nodes
list indicating which nodes are to be included in the calculation. By default, all nodes are included.

gmode
string indicating the type of graph being evaluated. "digraph" indicates that edges should be interpreted as directed; "graph" indicates that edges are undirected. This is currently ignored.

diag
boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

cmode
the rule to be used by symmetrize when symmetrizing dichotomous data; must be one of "weak" (for an OR rule), "strong" for an AND rule), "upper" (for a max rule), or "lower" (for a min rule). Set to "weak" by default, this parameter obviously has no effect on symmetric data.

tmaxdev
boolean indicating whether or not the theoretical maximum absolute deviation from the maximum nodal centrality should be returned. By default, tmaxdev==FALSE.

rescale
if true, centrality scores are rescaled such that they sum to 1.

tol
tolerance for near-singularities during matrix inversion (see solve).

Details

Actor information centrality is a hybrid measure which relates to both path-length indices (e.g., closeness, graph centrality) and to walk-based eigenmeasures (e.g., eigenvector centrality, Bonacich power). In particular, the information centrality of a given actor can be understood to be the harmonic average of the “bandwidth” for all paths originating with said individual (where the bandwidth is taken to be inversely related to path length). Formally, the index is constructed as follows. First, we take G to be an undirected (but possibly valued) graph – symmetrizing if necessary – with (possibly valued) adjacency matrix A. From this, we remove all isolates (whose information centralities are zero in any event) and proceed to create the weighted connection matrix

C = B^-1

where B is a pseudo-adjacency matrix formed by replacing the diagonal of 1-A with one plus each actor's degree. Given the above, let T be the trace of C with sum S_T, and let S_R be an arbitrary row sum (all rows of C have the same sum). The information centrality scores are then equal to

C_I = ( T + (S_T-2S_R)/|V(G)| )^-1

(recalling that the scores for any omitted vertices are 0).

In general, actors with higher information centrality are predicted to have greater control over the flow of information within a network; highly information-central individuals tend to have a large number of short paths to many others within the social structure. Because the raw centrality values can be difficult to interpret directly, rescaled values are sometimes preferred (see the rescale option). Though the use of path weights suggest information centrality as a possible replacement for closeness, the problem of inverting the B matrix poses problems of its own; as with all such measures, caution is advised on disconnected or degenerate structures.

Value

A vector, matrix, or list containing the centrality scores (depending on the number and size of the input graphs).

Note

The theoretical maximum deviation used here is not obtained with the star network; rather, the maximum occurs for an empty graph with one complete dyad, which is the model used here.

Author(s)

David Barron david.barron@jesus.ox.ac.uk

Carter T. Butts buttsc@uci.edu

References

Stephenson, K., and Zelen, M. (1989). “Rethinking Centrality: Methods and Applications.” Social Networks, 11, 1-37.

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

evcent, bonpow, closeness, graphcent, centralization

Examples

#Generate some test data
dat<-rgraph(10,mode="graph")
#Compute information centrality scores
infocent(dat)

[Package sna version 2.8 Index]