is.isolate {sna}	R Documentation

Is Ego an Isolate?

Description

Returns TRUE iff ego is an isolate in graph g of dat.

Usage

is.isolate(dat, ego, g=1, diag=FALSE)

Arguments

dat
one or more input graphs.

ego
index of the vertex (or a vector of vertices) to check.

g
which graph(s) should be examined?

diag
boolean indicating whether adjacency matrix diagonals (i.e., loops) contain meaningful data.

Details

In the valued case, any non-zero edge value is taken as sufficient to establish a tie.

Value

A boolean value (or vector thereof) indicating isolate status

Author(s)

Carter T. Butts buttsc@uci.edu

References

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

West, D.B. (1996). Introduction to Graph Theory. Upper Saddle River, NJ: Prentice Hall.

See Also

isolates, add.isolates

Examples

#Generate a test graph
g<-rgraph(20)
g[,4]<-0          #Create an isolate
g[4,]<-0

#Check for isolates
is.isolate(g,2)  #2 is almost surely not an isolate
is.isolate(g,4)  #4 is, by construction

[Package sna version 2.8 Index]