isolates {sna}	R Documentation

List the Isolates in a Graph or Graph Stack

Description

Returns a list of the isolates in the graph or graph set given by dat.

Usage

isolates(dat, diag=FALSE)

Arguments

dat
one or more input graphs.

diag
boolean indicating whether adjacency matrix diagonals (i.e., loops) contain meaningful data.

Value

A vector containing the isolates, or a list of vectors if more than one graph was specified

Author(s)

Carter T. Butts buttsc@uci.edu

References

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

West, D.B. (1996). Introduction to Graph Theory. Upper Saddle River, NJ: Prentice Hall.

See Also

is.isolate, add.isolates

Examples

#Generate a test graph
g<-rgraph(20)
g[,4]<-0          #Create an isolate
g[4,]<-0

#List the isolates
isolates(g)

[Package sna version 2.8 Index]