loadcent {sna}	R Documentation

Compute the Load Centrality Scores of Network Positions

Description

loadcent takes one or more graphs (dat) and returns the load centralities of positions (selected by nodes) within the graphs indicated by g. Depending on the specified mode, load on directed or undirected geodesics will be returned; this function is compatible with centralization, and will return the theoretical maximum absolute deviation (from maximum) conditional on size (which is used by centralization to normalize the observed centralization score).

Usage

loadcent(dat, g = 1, nodes = NULL, gmode = "digraph", diag = FALSE, 
    tmaxdev = FALSE, cmode = "directed", geodist.precomp = NULL, 
    rescale = FALSE, ignore.eval = TRUE)

Arguments

dat
one or more input graphs.

g
integer indicating the index of the graph for which centralities are to be calculated (or a vector thereof). By default, g=1.

nodes
vector indicating which nodes are to be included in the calculation. By default, all nodes are included.

gmode
string indicating the type of graph being evaluated. digraph indicates that edges should be interpreted as directed; graph indicates that edges are undirected. gmode is set to digraph by default.

diag
logical; should self-ties be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

tmaxdev
logical; return the theoretical maximum absolute deviation from the maximum nodal centrality (instead of the observed centrality scores)? By default, tmaxdev==FALSE.

cmode
string indicating the type of load centrality being computed (directed or undirected).

geodist.precomp
a geodist object precomputed for the graph to be analyzed (optional).

rescale
logical; if true, centrality scores are rescaled such that they sum to 1.

ignore.eval
logical; ignore edge values when computing shortest paths?

Details

Goh et al.'s load centrality (as reformulated by Brandes (2008)) is a betweenness-like measure defined through a hypothetical flow process. Specifically, it is assumed that each vertex sends a unit of some commodity to each other vertex to which it is connected (without edge or vertex capacity constraints), with routing based on a priority system: given an input of flow x arriving at vertex v with destination v', v divides x equally among all neigbors of minumum geodesic distance to the target. The total flow passing through a given v via this process is defined as v's load. Load is a potential alternative to betweenness for the analysis of flow structures operating well below their capacity constraints.

Value

A vector of centrality scores.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Brandes, U. (2008). “On Variants of Shortest-Path Betweenness Centrality and their Generic Computation.” Social Networks, 30, 136-145.

Goh, K.-I.; Kahng, B.; and Kim, D. (2001). “Universal Behavior of Load Distribution in Scale-free Networks.” Physical Review Letters, 87(27), 1-4.

See Also

betweenness

Examples

g<-rgraph(10)     #Draw a random graph with 10 members
loadcent(g)       #Compute load scores

[Package sna version 2.8 Index]