lubness {sna}	R Documentation

Compute Graph LUBness Scores

Description

lubness takes a graph set (dat) and returns the Krackhardt LUBness scores for the graphs selected by g.

Usage

lubness(dat, g=NULL)

Arguments

dat
one or more input graphs.

g
index values for the graphs to be utilized; by default, all graphs are selected.

Details

In the context of a directed graph G, two actors i and j may be said to have an upper bound iff there exists some actor k such that directed ki and kj paths belong to G. An upper bound l is known as a least upper bound for i and j iff it belongs to at least one ki and kj path (respectively) for all i,j upper bounds k; let L(i,j) be an indicator which returns 1 iff such an l exists, otherwise returning 0. Now, let G_1,G_2,...,G_n represent the weak components of G. For convenience, we denote the cardinalities of these graphs' vertex sets by |V(G)|=N and |V(G_i)|=N_i, for i in 1,...,n. Given this, the Krackhardt LUBness of G is given by

1-Sum(Sum(1-L(v_j,v_k),v_j,v_k in V(G_i)),i=1,...,n)/Sum((N_i-1)(N_i-2)/2,i=1,...,n)

Where all vertex pairs possess a least upper bound, Krackhardt's LUBness is equal to 1; in general, it approaches 0 as this condition is broached. (This convergence is problematic in certain cases due to the requirement that we sum violations across components; where a graph contains no components of size three or greater, Krackhardt's LUBness is not well-defined. lubness returns a NaN in these cases.)

LUBness is one of four measures (connectedness, efficiency, hierarchy, and lubness) suggested by Krackhardt for summarizing hierarchical structures. Each corresponds to one of four axioms which are necessary and sufficient for the structure in question to be an outtree; thus, the measures will be equal to 1 for a given graph iff that graph is an outtree. Deviations from unity can be interpreted in terms of failure to satisfy one or more of the outtree conditions, information which may be useful in classifying its structural properties.

Value

A vector of LUBness scores

Note

The four Krackhardt indices are, in general, nondegenerate for a relatively narrow band of size/density combinations (efficiency being the sole exception). This is primarily due to their dependence on the reachability graph, which tends to become complete rapidly as size/density increase. See Krackhardt (1994) for a useful simulation study.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Krackhardt, David. (1994). “Graph Theoretical Dimensions of Informal Organizations.” In K. M. Carley and M. J. Prietula (Eds.), Computational Organization Theory, 89-111. Hillsdale, NJ: Lawrence Erlbaum and Associates.

See Also

connectedness, efficiency, hierarchy, lubness, reachability

Examples

#Get LUBness scores for graphs of varying densities
lubness(rgraph(10,5,tprob=c(0.1,0.25,0.5,0.75,0.9)))

[Package sna version 2.8 Index]