mutuality {sna}	R Documentation

Find the Mutuality of a Graph

Description

Returns the mutuality scores of the graphs indicated by g in dat.

Usage

mutuality(dat, g=NULL)

Arguments

dat
one or more input graphs.

g
a vector indicating which elements of dat should be analyzed; by default, all graphs are included.

Details

The mutuality of a digraph G is defined as the number of complete dyads (i.e., i<->j) within G. (Compare this to dyadic reciprocity, the fraction of dyads within G which are symmetric.) Mutuality is commonly employed as a measure of reciprocal tendency within the p* literature; although mutuality can be very hard to interpret in practice, it is much better behaved than many alternative measures.

Value

One or more mutuality scores

Author(s)

Carter T. Butts buttsc@uci.edu

References

Moreno, J.L., and Jennings, H.H. (1938). “Statistics of Social Configurations.” Sociometry, 1, 342-374.

See Also

grecip

Examples

#Create some random graphs
g<-rgraph(15,3)

#Get mutuality and reciprocity scores
mutuality(g)
grecip(g)         #Compare with mutuality

[Package sna version 2.8 Index]