nties {sna}	R Documentation

Find the Number of Possible Ties in a Given Graph or Graph Stack

Description

nties returns the number of possible edges in each element of dat, given mode and diag.

Usage

nties(dat, mode="digraph", diag=FALSE)

Arguments

dat
a graph or set thereof.

mode
one of “digraph”, “graph”, and “hgraph”.

diag
a boolean indicating whether or not diagonal entries (loops) should be treated as valid data; ignored for hypergraphic (“hgraph”) data.

Details

nties is used primarily to automate maximum edge counts for use with normalization routines.

Value

The number of possible edges, or a vector of the same

Note

For two-mode (hypergraphic) data, the value returned isn't technically the number of edges per se, but rather the number of edge memberships.

Author(s)

Carter T. Butts buttsc@uci.edu

Examples

#How many possible edges in a loopless digraph of order 15?
nties(rgraph(15),diag=FALSE)

[Package sna version 2.8 Index]