plot.blockmodel {sna}	R Documentation

Plotting for blockmodel Objects

Description

Displays a plot of the blocked data matrix, given a blockmodel object.

Usage

## S3 method for class 'blockmodel'
plot(x, ...)

Arguments

x
An object of class blockmodel

...
Further arguments passed to or from other methods

Details

Plots of the blocked data matrix (i.e., the data matrix with rows and columns permuted to match block membership) can be useful in assessing the strength of the block solution (particularly for clique detection and/or regular equivalence).

Value

None

Author(s)

Carter T. Butts buttsc@uci.edu

References

White, H.C.; Boorman, S.A.; and Breiger, R.L. (1976). “Social Structure from Multiple Networks I: Blockmodels of Roles and Positions.” American Journal of Sociology, 81, 730-779.

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

blockmodel, plot.sociomatrix

Examples

#Create a random graph with _some_ edge structure
g.p<-sapply(runif(20,0,1),rep,20)  #Create a matrix of edge 
                                   #probabilities
g<-rgraph(20,tprob=g.p)            #Draw from a Bernoulli graph 
                                   #distribution

#Cluster based on structural equivalence
eq<-equiv.clust(g)

#Form a blockmodel with distance relaxation of 10
b<-blockmodel(g,eq,h=10)
plot(b)                            #Plot it

[Package sna version 2.8 Index]