plot.cugtest {sna}	R Documentation

Plotting for cugtest Objects

Description

Plots the distribution of a CUG test statistic.

Usage

## S3 method for class 'cugtest'
plot(x, mode="density", ...)

Arguments

x
A cugtest object

mode
“density” for kernel density estimation, “hist” for histogram

...
Additional arguments to plot

Details

In addition to the quantiles associated with a CUG test, it is often useful to examine the form of the distribution of the test statistic. plot.cugtest facilitates this.

Value

None

Author(s)

Carter T. Butts buttsc@uci.edu

References

Anderson, B.S.; Butts, C.T.; and Carley, K.M. (1999). “The Interaction of Size and Density with Graph-Level Indices.” Social Networks, 21(3), 239-267.

See Also

cugtest

Examples

#Draw two random graphs, with different tie probabilities
dat<-rgraph(20,2,tprob=c(0.2,0.8))

#Is their correlation higher than would be expected, conditioning 
#only on size?
cug<-cugtest(dat,gcor,cmode="order")
summary(cug)
plot(cug)

#Now, let's try conditioning on density as well.
cug<-cugtest(dat,gcor)
plot(cug)

[Package sna version 2.8 Index]