plot.equiv.clust {sna}	R Documentation

Plot an equiv.clust Object

Description

Plots a hierarchical clustering of node positions as generated by equiv.clust.

Usage

## S3 method for class 'equiv.clust'
plot(x, labels=NULL, ...)

Arguments

x
An equiv.clust object

labels
A vector of vertex labels

...
Additional arguments to plot.hclust

Details

plot.equiv.clust is actually a front-end to plot.hclust; see the latter for more additional documentation.

Value

None.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Breiger, R.L.; Boorman, S.A.; and Arabie, P. (1975). “An Algorithm for Clustering Relational Data with Applications to Social Network Analysis and Comparison with Multidimensional Scaling.” Journal of Mathematical Psychology, 12, 328-383.

Burt, R.S. (1976). “Positions in Networks.” Social Forces, 55, 93-122.

Wasserman, S., and Faust, K. Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

equiv.clust, plot.hclust

Examples

#Create a random graph with _some_ edge structure
g.p<-sapply(runif(20,0,1),rep,20)  #Create a matrix of edge 
                                   #probabilities
g<-rgraph(20,tprob=g.p)            #Draw from a Bernoulli graph 
                                   #distribution

#Cluster based on structural equivalence
eq<-equiv.clust(g)
plot(eq)

[Package sna version 2.8 Index]