plot.qaptest {sna}	R Documentation

Plotting for qaptest Objects

Description

Plots the Distribution of a QAP Test Statistic.

Usage

## S3 method for class 'qaptest'
plot(x, mode="density", ...)

Arguments

x
A qaptest object

mode
“density” for kernel density estimation, “hist” for histogram

...
Additional arguments to plot

Details

In addition to the quantiles associated with a QAP test, it is often useful to examine the form of the distribution of the test statistic. plot.qaptest facilitates this.

Value

None

Author(s)

Carter T. Butts buttsc@uci.edu

References

Hubert, L.J., and Arabie, P. (1989). “Combinatorial Data Analysis: Confirmatory Comparisons Between Sets of Matrices.” Applied Stochastic Models and Data Analysis, 5, 273-325.

Krackhardt, D. (1987). “QAP Partialling as a Test of Spuriousness.” Social Networks, 9 171-186.

Krackhardt, D. (1988). “Predicting With Networks: Nonparametric Multiple Regression Analyses of Dyadic Data.” Social Networks, 10, 359-382.

See Also

qaptest

Examples

#Generate three graphs
g<-array(dim=c(3,10,10))
g[1,,]<-rgraph(10)
g[2,,]<-rgraph(10,tprob=g[1,,]*0.8)
g[3,,]<-1; g[3,1,2]<-0              #This is nearly a clique

#Perform qap tests of graph correlation
q.12<-qaptest(g,gcor,g1=1,g2=2)
q.13<-qaptest(g,gcor,g1=1,g2=3)

#Examine the results
summary(q.12)
plot(q.12)
summary(q.13)
plot(q.13)

[Package sna version 2.8 Index]