prestige {sna}	R Documentation

Calculate the Vertex Prestige Scores

Description

prestige takes one or more graphs (dat) and returns the prestige scores of positions (selected by nodes) within the graphs indicated by g. Depending on the specified mode, prestige based on any one of a number of different definitions will be returned. This function is compatible with centralization, and will return the theoretical maximum absolute deviation (from maximum) conditional on size (which is used by centralization to normalize the observed centralization score).

Usage

prestige(dat, g=1, nodes=NULL, gmode="digraph", diag=FALSE, 
    cmode="indegree", tmaxdev=FALSE, rescale=FALSE, tol=1e-07)

Arguments

dat
one or more input graphs.

g
integer indicating the index of the graph for which centralities are to be calculated (or a vector thereof). By default, g==1.

nodes
vector indicating which nodes are to be included in the calculation. By default, all nodes are included.

gmode
string indicating the type of graph being evaluated. "digraph" indicates that edges should be interpreted as directed; "graph" indicates that edges are undirected. gmode is set to "digraph" by default.

diag
boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

cmode
one of "indegree", "indegree.rownorm", "indegree.rowcolnorm", "eigenvector", "eigenvector.rownorm", "eigenvector.colnorm", "eigenvector.rowcolnorm", "domain", or "domain.proximity".

tmaxdev
boolean indicating whether or not the theoretical maximum absolute deviation from the maximum nodal centrality should be returned. By default, tmaxdev==FALSE.

rescale
if true, centrality scores are rescaled such that they sum to 1.

tol
Currently ignored

Details

"Prestige" is the name collectively given to a range of centrality scores which focus on the extent to which one is nominated by others. The definitions supported here are as follows:

indegree: indegree centrality

indegree.rownorm: indegree within the row-normalized graph

indegree.rowcolnorm: indegree within the row-column normalized graph

eigenvector: eigenvector centrality within the transposed graph (i.e., incoming ties recursively determine prestige)

eigenvector.rownorm: eigenvector centrality within the transposed row-normalized graph

eigenvector.colnorm: eigenvector centrality within the transposed column-normalized graph

eigenvector.rowcolnorm: eigenvector centrality within the transposed row/column-normalized graph

domain: indegree within the reachability graph (Lin's unweighted measure)

domain.proximity: Lin's proximity-weighted domain prestige

Note that the centralization of prestige is simply the extent to which one actor has substantially greater prestige than others; the underlying definition is the same.

Value

A vector, matrix, or list containing the prestige scores (depending on the number and size of the input graphs).

Warning

Making adjacency matrices doubly stochastic (row-column normalization) is not guaranteed to work. In general, be wary of attempting to try normalizations on graphs with degenerate rows and columns.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Lin, N. (1976). Foundations of Social Research. New York: McGraw Hill.

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

centralization

Examples

g<-rgraph(10)                 #Draw a random graph with 10 members
prestige(g,cmode="domain")    #Compute domain prestige scores

[Package sna version 2.8 Index]