read.dot {sna}	R Documentation

Read Graphviz DOT Files

Description

Reads network information in Graphviz's DOT format, returning an adjacency matrix.

Usage

read.dot(...)

Arguments

...
The name of the file whence to import the data, or else a connection object (suitable for processing by readLines.

Details

The Graphviz project's DOT language is a simple but flexible tool for describing graphs. See the included reference for details.

Value

The imported graph, in adjacency matrix form.

Author(s)

Matthijs den Besten matthijs.denbesten@gmail.com

References

Graphviz Project. "The DOT Language." http://www.graphviz.org/doc/info/lang.html

See Also

read.nos, write.nos, write.dl

[Package sna version 2.8 Index]