rguman {sna}	R Documentation

Draw Dyad Census-Conditioned Random Graphs

Description

rguman generates random draws from a dyad census-conditioned uniform random graph distribution.

Usage

rguman(n, nv, mut = 0.25, asym = 0.5, null = 0.25, 
    method = c("probability", "exact"), return.as.edgelist = FALSE)

Arguments

n
the number of graphs to generate.

nv
the size of the vertex set (|V(G)|) for the random graphs.

mut
if method=="probability", the probability of obtaining a mutual dyad; otherwise, the number of mutual dyads.

asym
if method=="probability", the probability of obtaining an asymmetric dyad; otherwise, the number of asymmetric dyads.

null
if method=="probability", the probability of obtaining a null dyad; otherwise, the number of null dyads.

method
the generation method to use. "probability" results in a multinomial dyad distribution (conditional on the underlying rates), while "exact" results in a uniform draw conditional on the exact dyad distribution.

return.as.edgelist
logical; should the resulting graphs be returned in edgelist form?

Details

A simple generalization of the Erdos-Renyi family, the U|MAN distributions are uniform on the set of graphs, conditional on order (size) and the dyad census. As with the E-R case, there are two U|MAN variants. The first (corresponding to method=="probability") takes dyad states as independent multinomials with parameters m (for mutuals), a (for asymmetrics), and n (for nulls). The resulting pmf is then

p(G=g|m,a,n) = (M+A+N)!/(M!A!N!) m^M a^A n^N,

where M, A, and N are realized counts of mutual, asymmetric, and null dyads, respectively. (See dyad.census for an explication of dyad types.)

The second U|MAN variant is selected by method=="exact", and places equal mass on all graphs having the specified (exact) dyad census. The corresponding pmf is

p(G=g|M,A,N) = M!A!N!/(M+A+N)!.

U|MAN graphs provide a natural baseline model for networks which are constrained by size, density, and reciprocity. In this way, they provide a bridge between edgewise models (e.g., the E-R family) and models with higher order dependence (e.g., the Markov graphs).

Value

A matrix or array containing the drawn adjacency matrices

Note

The famous mathematicians referenced in this man page now have misspelled names, due to R's difficulty with accent marks.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Holland, P.W. and Leinhardt, S. (1976). “Local Structure in Social Networks.” In D. Heise (Ed.), Sociological Methodology, pp 1-45. San Francisco: Jossey-Bass.

Wasserman, S. and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

See Also

rgraph, rgnm, dyad.census

Examples

#Show some examples of extreme U|MAN graphs
gplot(rguman(1,10,mut=45,asym=0,null=0,method="exact")) #Clique
gplot(rguman(1,10,mut=0,asym=45,null=0,method="exact")) #Tournament
gplot(rguman(1,10,mut=0,asym=0,null=45,method="exact")) #Empty

#Draw a sample of multinomial U|MAN graphs
g<-rguman(5,10,mut=0.15,asym=0.05,null=0.8)

#Examine the dyad census
dyad.census(g)

[Package sna version 2.8 Index]