rmperm {sna}	R Documentation

Randomly Permute the Rows and Columns of an Input Matrix

Description

Given an input matrix (or stack thereof), rmperm performs a (random) simultaneous row/column permutation of the input data.

Usage

rmperm(m)

Arguments

m
a matrix, or stack thereof (or a graph set, for that matter).

Details

Random matrix permutations are the essence of the QAP test; see qaptest for details.

Value

The permuted matrix (or matrices)

Author(s)

Carter T. Butts buttsc@uci.edu

See Also

rperm

Examples

#Generate an input matrix
g<-rgraph(5)
g             #Examine it

#Examine a random permutation
rmperm(g)

[Package sna version 2.8 Index]