sna.operators {sna}	R Documentation

Graphical Operators

Description

These operators allow for algebraic manupulation of graph adjacency matrices.

Usage

## S3 method for class 'matrix'
e1 %c% e2

Arguments

e1
an (unvalued) adjacency matrix.

e2
another (unvalued) adjacency matrix.

Details

Currently, only one operator is supported. x %c% y returns the adjacency matrix of the composition of graphs with adjacency matrices x and y (respectively). (Note that this may contain loops.)

Value

The resulting adjacency matrix.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Wasserman, S. and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: University of Cambridge Press.

Examples

#Create an in-star
g<-matrix(0,6,6)
g[2:6,1]<-1
gplot(g)

#Compose g with its transpose
gcgt<-g%c%t(g)
gplot(gcgt,diag=TRUE)
gcgt

[Package sna version 2.8 Index]