sr2css {sna}	R Documentation

Convert a Row-wise Self-Report Matrix to a CSS Matrix with Missing Observations

Description

Given a matrix in which the ith row corresponds to i's reported relations, sr2css creates a graph stack in which each element represents a CSS slice with missing observations.

Usage

sr2css(net)

Arguments

net
an adjacency matrix.

Details

A cognitive social structure (CSS) is an nxnxn array in which the ith matrix corresponds to the ith actor's perception of the entire network. Here, we take a conventional self-report data structure and put it in CSS format for routines (such as bbnam) which require this.

Value

An array (graph stack) containing the CSS

Note

A row-wise self-report matrix doesn't contain a great deal of data, and the data in question is certainly not an ignorable sample of the individual's CSS for most purposes. The provision of this routine should not be perceived as license to substitute SR for CSS data at will.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Krackhardt, D. (1987). Cognitive Social Structures, 9, 109-134.

Examples

#Start with some random reports
g<-rgraph(10)

#Transform to CSS format
c<-sr2css(g)

[Package sna version 2.8 Index]