stresscent {sna}	R Documentation

Compute the Stress Centrality Scores of Network Positions

Description

stresscent takes one or more graphs (dat) and returns the stress centralities of positions (selected by nodes) within the graphs indicated by g. Depending on the specified mode, stress on directed or undirected geodesics will be returned; this function is compatible with centralization, and will return the theoretical maximum absolute deviation (from maximum) conditional on size (which is used by centralization to normalize the observed centralization score).

Usage

stresscent(dat, g=1, nodes=NULL, gmode="digraph", 
    diag=FALSE, tmaxdev=FALSE, cmode="directed", 
    geodist.precomp=NULL, rescale=FALSE, ignore.eval=TRUE)

Arguments

dat
one or more input graphs.

g
Integer indicating the index of the graph for which centralities are to be calculated (or a vector thereof). By default, g==1.

nodes
list indicating which nodes are to be included in the calculation. By default, all nodes are included.

gmode
string indicating the type of graph being evaluated. "digraph" indicates that edges should be interpreted as directed; "graph" indicates that edges are undirected. gmode is set to "digraph" by default.

diag
boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

tmaxdev
boolean indicating whether or not the theoretical maximum absolute deviation from the maximum nodal centrality should be returned. By default, tmaxdev==FALSE.

cmode
string indicating the type of betweenness centrality being computed (directed or undirected geodesics).

geodist.precomp
a geodist object precomputed for the graph to be analyzed (optional).

rescale
if true, centrality scores are rescaled such that they sum to 1.

ignore.eval
logical; should edge values be ignored when calculating density?

Details

The stress of a vertex, v, is given by

C_S(v) = sum( g_ivj, i,j: i!=j,i!=v,j!=v)

where g_ijk is the number of geodesics from i to k through j. Conceptually, high-stress vertices lie on a large number of shortest paths between other vertices; they can thus be thought of as “bridges” or “boundary spanners.” Compare this with betweenness, which weights shortest paths by the inverse of their redundancy.

Value

A vector, matrix, or list containing the centrality scores (depending on the number and size of the input graphs).

Note

Judicious use of geodist.precomp can save a great deal of time when computing multiple path-based indices on the same network.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Shimbel, A. (1953). “Structural Parameters of Communication Networks.” Bulletin of Mathematical Biophysics, 15:501-507.

See Also

centralization

Examples

g<-rgraph(10)     #Draw a random graph with 10 members
stresscent(g)     #Compute stress scores

[Package sna version 2.8 Index]