structure.statistics {sna}	R Documentation

Compute Network Structure Statistics

Description

Computes the structure statistics for the graph(s) in dat.

Usage

structure.statistics(dat, geodist.precomp = NULL)

Arguments

dat
one or more input graphs.

geodist.precomp
a geodist object (optional).

Details

Let G=(V,E) be a graph of order N, and let d(i,j) be the geodesic distance from vertex i to vertex j in G. The "structure statistics" of G are then given by the series s_0,…,s_{N-1}, where s_i = sum(sum(I(d(j,k)<=i, k in V), j in V)/N^2 and I is the standard indicator function. Intuitively, s_i is the expected fraction of G which lies within distance i of a randomly chosen vertex. As such, the structure statistics provide an index of global connectivity.

Structure statistics have been of particular importance to biased net theorists, because of the link with Rapoport's original tracing model. They may also be used along with component distributions or connectedness scores as descriptive indices of connectivity at the graph-level.

Value

A vector, matrix, or list (depending on dat) containing the structure statistics.

Note

The term "structure statistics" has been used somewhat loosely in the literature, a trend which seems to be accelerating. Users should carefully check references before comparing results generated by this routine with those appearing in published work.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Fararo, T.J. (1981). “Biased networks and social structure theorems. Part I.” Social Networks, 3, 137-159.

Fararo, T.J. (1984). “Biased networks and social structure theorems. Part II.” Social Networks, 6, 223-258.

Fararo, T.J. and Sunshine, M.H. (1964). “A study of a biased friendship net.” Syracuse, NY: Youth Development Center.

See Also

geodist, component.dist, connectedness, bn

Examples

#Generate a moderately sparse Bernoulli graph
g<-rgraph(100,tp=1.5/99)

#Compute the structure statistics for g
ss<-structure.statistics(g)
plot(0:99,ss,xlab="Mean Coverage",ylab="Distance")

[Package sna version 2.8 Index]