symmetrize {sna}	R Documentation

Symmetrize an Adjacency Matrix

Description

Symmetrizes the elements of mats according to the rule in rule.

Usage

symmetrize(mats, rule="weak", return.as.edgelist=FALSE)

Arguments

mats
a graph or graph stack.

rule
one of “upper”, “lower”, “strong” or “weak”.

return.as.edgelist
logical; should the symmetrized graphs be returned in edgelist form?

Details

The rules used by symmetrize are as follows:

upper: Copy the upper triangle over the lower triangle

lower: Copy the lower triangle over the upper triangle

strong: i<->j iff i->j and i<-j (AND rule)

weak: i<->j iff i->j or i<-j (OR rule)

Value

The symmetrized graph stack

Author(s)

Carter T. Butts buttsc@uci.edu

References

Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.

Examples

#Generate a graph
g<-rgraph(5)

#Weak symmetrization
symmetrize(g)

#Strong symmetrization
symmetrize(g,rule="strong")

[Package sna version 2.8 Index]