triad.census {sna}	R Documentation

Compute the Davis and Leinhardt Triad Census

Description

triad.census returns the Davis and Leinhardt triad census of the elements of dat indicated by g.

Usage

triad.census(dat, g=NULL, mode = c("digraph", "graph"))

Arguments

dat
a graph or graph stack.

g
the elements of dat to process.

mode
string indicating the directedness of edges; "digraph" implies a directed structure, whereas "graph" implies an undirected structure.

Details

The Davis and Leinhardt triad census consists of a classification of all directed triads into one of 16 different categories; the resulting distribution can be compared against various null models to test for the presence of configural biases (e.g., transitivity bias). triad.census is a front end for the triad.classify routine, performing the classification for all triads within the selected graphs. The results are placed in the order indicated by the column names; this is the same order as presented in the triad.classify documentation, to which the reader is referred for additional details.

In the undirected case, the triad census reduces to four states (based on the number of edges in each triad. Where mode=="graph", this is returned instead.

Compare triad.census to dyad.census, the dyadic equivalent.

Value

A matrix whose 16 columns contain the counts of triads by class for each graph, in the directed case. In the undirected case, only 4 columns are used.

Warning

Valued data may cause strange behavior with this routine. Dichotomize the data first.

Author(s)

Carter T. Butts buttsc@uci.edu

References

Davis, J.A. and Leinhardt, S. (1972). “The Structure of Positive Interpersonal Relations in Small Groups.” In J. Berger (Ed.), Sociological Theories in Progress, Volume 2, 218-251. Boston: Houghton Mifflin.

Wasserman, S., and Faust, K. (1994). “Social Network Analysis: Methods and Applications.” Cambridge: Cambridge University Press.

See Also

triad.classify, dyad.census, kcycle.census, kpath.census, gtrans

Examples

#Generate a triad census of random data with varying densities
triad.census(rgraph(15,5,tprob=c(0.1,0.25,0.5,0.75,0.9)))

[Package sna version 2.8 Index]